requirements.txt
asgiref==3.8.1
certifi==2024.2.2
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
filelock==3.13.3
gunicorn==21.2.0
Markdown==3.6
mysql-connector-python==8.3.0
packaging==24.0
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2==2.9.9
setuptools==69.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
virtualenv==20.25.1
requests==2.26.0
django-rest-passwordreset==1.4.1
